se arreglaron las operaciones de alu.c en cuanto a la modificacion de banderas ya que segun el manual del arm cortex m0 el gunas funciones de la no modifican ciertas banderas. se organizaron las instrucciones de salto. Se implementan las funciones de decoder para reconocer las instrucciones dentro del archivo code.txt verificando el funcionamiento con el codigo ya escrito en code.txt

+++ registros/screen.h
@@ -8,6 +8,6 @@
 
 #include "colors.h"
 
-void showRegisters(unsigned long int *registers, size_t len);
+void showRegisters(uint32_t *registers, size_t len);
 
 #endif	/* _SCREEN_H_ */
